+++ src/core.ts
@@ -48,7 +48,8 @@ export {
 } from './rx-document-prototype-merge';
 
 export {
-    isInstanceOf as isRxQuery
+    isInstanceOf as isRxQuery,
+    normalizeMangoQuery
 } from './rx-query';
 
 export {