+++ src/plugins/dev-mode/check-query.ts
@@ -1,5 +1,10 @@
-import { RxPluginPreCreateRxQueryArgs, MangoQuery } from '../../types';
-import { newRxTypeError } from '../../rx-error';
+import type {
+    RxPluginPreCreateRxQueryArgs,
+    MangoQuery,
+    RxPluginPrePrepareQueryArgs
+} from '../../types';
+import deepEqual from 'fast-deep-equal';
+import { newRxError, newRxTypeError } from '../../rx-error';
 
 /**
  * accidentially passing a non-valid object into the query params
@@ -20,7 +25,8 @@ export function checkQuery(args: RxPluginPreCreateRxQueryArgs) {
         'selector',
         'limit',
         'skip',
-        'sort'
+        'sort',
+        'index'
     ];
     Object.keys(args.queryObj).forEach(key => {
         if (!(validKeys as string[]).includes(key)) {
@@ -28,8 +34,35 @@ export function checkQuery(args: RxPluginPreCreateRxQueryArgs) {
                 op: args.op,
                 collection: args.collection.name,
                 queryObj: args.queryObj,
-                key
+                key,
+                args: {
+                    validKeys
+                }
             });
         }
     });
+}
+
+
+export function checkMangoQuery(args: RxPluginPrePrepareQueryArgs) {
+    /**
+     * ensure if custom index is set,
+     * it is also defined in the schema
+     */
+    const schema = args.rxQuery.collection.schema.normalized;
+    const schemaIndexes = schema.indexes ? schema.indexes : [];
+    const index = args.mangoQuery.index;
+    if (index) {
+        const isInSchema = schemaIndexes.find(schemaIndex => deepEqual(schemaIndex, index));
+        if (!isInSchema) {
+            throw newRxError(
+                'QU12',
+                {
+                    collection: args.rxQuery.collection.name,
+                    query: args.mangoQuery,
+                    schema
+                }
+            );
+        }
+    }
 }